+++ jaeseung/src/dependency.rs
@@ -1,0 +1,1 @@
+pub static USER_REPO: repo::UserFileRepository = UserFileRepository {};
