+++ jaeseung/src/user/user.rs
@@ -1,0 +1,4 @@
+pub struct User {
+    pub id: String,
+    pub password: String
+}
